+++ CHANGELOG.rst
@@ -36,6 +36,7 @@ Changed
     - Improve characteristics list parsing (#616, #618)
     - Always import ``material_type`` field for ``GenericMaterial``
     - Do not replace title or description in ``Investigation`` if not provided
+    - Display configuration in Overview as badge
 
 Fixed
 -----
@@ -47,6 +48,7 @@ Fixed
 - **Samplesheets**
     - "Sequence item 1" render error manifesting with BII-I-1 example (#620)
     - Redundant unit/value parsing for comments during import (#629)
+    - Missing label for unknown configuration in Overview (#638)
 
 Removed
 -------